+++ IB-Tracker/src/Utilities/Colors.h
@@ -26,32 +26,32 @@ SOFTWARE.
 
 #pragma once
 
-const DWORD COLOR_BLACK = Color::Black;
-const DWORD COLOR_GRAYDARK = Color::MakeARGB(255, 22, 26, 27);
-const DWORD COLOR_GRAYMEDIUM = Color::MakeARGB(255, 51, 51, 51);
-const DWORD COLOR_GRAYLIGHT = Color::MakeARGB(255, 68, 68, 68);
+const DWORD COLOR_BLACK = (DWORD) Color::Black;
+const DWORD COLOR_GRAYDARK = (DWORD) Color::MakeARGB(255, 22, 26, 27);
+const DWORD COLOR_GRAYMEDIUM = (DWORD) Color::MakeARGB(255, 51, 51, 51);
+const DWORD COLOR_GRAYLIGHT = (DWORD) Color::MakeARGB(255, 68, 68, 68);
 
-const DWORD COLOR_WHITE = Color::White;
-const DWORD COLOR_WHITEDARK = Color::Gray;
-const DWORD COLOR_WHITEMEDIUM = Color::MakeARGB(255, 157, 165, 180);
-const DWORD COLOR_WHITELIGHT = Color::MakeARGB(255, 212, 212, 212);
+const DWORD COLOR_WHITE = (DWORD) Color::White;
+const DWORD COLOR_WHITEDARK = (DWORD) Color::Gray;
+const DWORD COLOR_WHITEMEDIUM = (DWORD) Color::MakeARGB(255, 157, 165, 180);
+const DWORD COLOR_WHITELIGHT = (DWORD) Color::MakeARGB(255, 212, 212, 212);
 
-const DWORD COLOR_MAGENTA = Color::Magenta;
-const DWORD COLOR_YELLOW = Color::Yellow;
-const DWORD COLOR_GREEN = Color::MakeARGB(255, 72, 151, 13);
-const DWORD COLOR_RED = Color::Firebrick;
-const DWORD COLOR_ORANGE = Color::MakeARGB(255, 193, 98, 24);  // burnt orange
-const DWORD COLOR_PURPLE = Color::MediumOrchid;
-const DWORD COLOR_BLUE = Color::CornflowerBlue;
-const DWORD COLOR_PINK = Color::Pink;
-const DWORD COLOR_TEAL = Color::Teal;
-const DWORD COLOR_KHAKI = Color::DarkKhaki;
+const DWORD COLOR_MAGENTA = (DWORD) Color::Magenta;
+const DWORD COLOR_YELLOW = (DWORD) Color::Yellow;
+const DWORD COLOR_GREEN = (DWORD) Color::MakeARGB(255, 72, 151, 13);
+const DWORD COLOR_RED = (DWORD) Color::Firebrick;
+const DWORD COLOR_ORANGE = (DWORD) Color::MakeARGB(255, 193, 98, 24);  // burnt orange
+const DWORD COLOR_PURPLE = (DWORD) Color::MediumOrchid;
+const DWORD COLOR_BLUE = (DWORD) Color::CornflowerBlue;
+const DWORD COLOR_PINK = (DWORD) Color::Pink;
+const DWORD COLOR_TEAL = (DWORD) Color::Teal;
+const DWORD COLOR_KHAKI = (DWORD) Color::DarkKhaki;
 
-const DWORD COLOR_MENUNOTCH = Color::MakeARGB(255, 22, 26, 27);
-const DWORD COLOR_SEPARATOR = Color::MakeARGB(255, 53, 59, 69);
-const DWORD COLOR_SELECTION = Color::MakeARGB(255, 44, 49, 58);
+const DWORD COLOR_MENUNOTCH = (DWORD) Color::MakeARGB(255, 22, 26, 27);
+const DWORD COLOR_SEPARATOR = (DWORD) Color::MakeARGB(255, 53, 59, 69);
+const DWORD COLOR_SELECTION = (DWORD) Color::MakeARGB(255, 44, 49, 58);
 
-const DWORD COLOR_SCROLLBARBACK = Color::MakeARGB(255, 22, 26, 27);
-const DWORD COLOR_SCROLLBARDIVIDER = Color::MakeARGB(255, 64, 67, 73);
-const DWORD COLOR_SCROLLBARTHUMB = Color::MakeARGB(255, 51, 51, 51);
+const DWORD COLOR_SCROLLBARBACK = (DWORD) Color::MakeARGB(255, 22, 26, 27);
+const DWORD COLOR_SCROLLBARDIVIDER = (DWORD) Color::MakeARGB(255, 64, 67, 73);
+const DWORD COLOR_SCROLLBARTHUMB = (DWORD) Color::MakeARGB(255, 51, 51, 51);
